Make specialized functions for uint16 and uint32. Add algorithm 6 support. Pass the parameters.

+++ eiger-zmq/boost_python/config_manager.py
@@ -25,7 +25,7 @@ cheetah {
  MaxPixCount = 40
   .type = int
  LocalBGRadius = 2
-  .type = float
+  .type = int
  MinPeakSeparation = 0
   .type = float
 }
